+++ src/presenter/main-presenter.js
@@ -1,6 +1,6 @@
 import { render } from '../render';
-import AddPointView from '../view/add-point-view';
-// import EditPointView from '../view/edit-point-view';
+import AddFormView from '../view/add-form-view';
+// import EditFormView from '../view/edit-point-view';
 import FiltersView from '../view/filters-view';
 import PointView from '../view/point-view';
 import PointsListView from '../view/points-list-view';
@@ -20,7 +20,7 @@ export default class MainPresenter {
   init() {
     render(new FiltersView(), this.filtersContainer);
     render(new SortView(), this.tripEventsContainer);
-    render(new AddPointView(), this.tripEventsContainer);
+    render(new AddFormView(), this.tripEventsContainer);
     render(this.pointsListElement, this.tripEventsContainer);
     POINTS.map((point) => render(new PointView(point), this.pointsListElement.getElement()));
   }
