+++ src/view/edit-form-view.js
@@ -1,14 +1,13 @@
-import { OFFERS, PHOTOS_PATHS } from '../data';
 import { createElement } from '../render';
+import { OFFERS } from '../data';
 import {
   createEventTypeListTemplate,
   createOffersListTemplate,
-  createPhotosContainerTemplate
-} from './add-and-edit-points-templates';
+} from './forms-templates';
 
-// $======================== AddPointView ========================$ //
+// $======================== EditFormView ========================$ //
 
-const createAddPointTemplate = () => /*html*/`
+const createEditPointTemplate = () => /*html*/`
   <form class="event event--edit" action="#" method="post">
     <header class="event__header">
       <div class="event__type-wrapper">
@@ -26,7 +25,7 @@ const createAddPointTemplate = () => /*html*/`
         <label class="event__label  event__type-output" for="event-destination-1">
           Flight
         </label>
-        <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="Geneva" list="destination-list-1">
+        <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="Chamonix" list="destination-list-1">
         <datalist id="destination-list-1">
           <option value="Amsterdam"></option>
           <option value="Geneva"></option>
@@ -36,10 +35,10 @@ const createAddPointTemplate = () => /*html*/`
 
       <div class="event__field-group  event__field-group--time">
         <label class="visually-hidden" for="event-start-time-1">From</label>
-        <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="19/03/19 00:00">
+        <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="18/03/19 12:25">
         &mdash;
         <label class="visually-hidden" for="event-end-time-1">To</label>
-        <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="19/03/19 00:00">
+        <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="18/03/19 13:35">
       </div>
 
       <div class="event__field-group  event__field-group--price">
@@ -51,8 +50,10 @@ const createAddPointTemplate = () => /*html*/`
       </div>
 
       <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-      <button class="event__reset-btn" type="reset">Cancel</button>
-
+      <button class="event__reset-btn" type="reset">Delete</button>
+      <button class="event__rollup-btn" type="button">
+        <span class="visually-hidden">Open event</span>
+      </button>
     </header>
 
     <section class="event__details">
@@ -66,19 +67,16 @@ const createAddPointTemplate = () => /*html*/`
       <section class="event__section  event__section--destination">
         <h3 class="event__section-title  event__section-title--destination">Destination</h3>
         <p class="event__destination-description">
-          Geneva is a city in Switzerland that lies at the southern tip of expansive Lac Léman (Lake Geneva). Surrounded by the Alps and Jura mountains, the city has views of dramatic Mont Blanc.
+          Chamonix-Mont-Blanc (usually shortened to Chamonix) is a resort area near the junction of France, Switzerland and Italy. At the base of Mont Blanc, the highest summit in the Alps, it's renowned for its skiing.
         </p>
-
-        ${createPhotosContainerTemplate(PHOTOS_PATHS)}
-
       </section>
     </section>
   </form>
 `;
 
-export default class AddPointView {
+export default class EditFormView {
   getTemplate() {
-    return createAddPointTemplate();
+    return createEditPointTemplate();
   }
 
   getElement() {
