+++ src/view/filters-view.js
@@ -3,7 +3,7 @@ import {FILTER_TYPES} from '../data';
 
 // $======================== FiltersView ========================$ //
 
-const createFilterItemTemplate = (filterType) => `
+const createFilterItemTemplate = (filterType) => /*html*/`
   <div class="trip-filters__filter">
     <input id="filter-${filterType}" class="trip-filters__filter-input visually-hidden" type="radio" name="trip-filter" value="${filterType}" />
       <label class="trip-filters__filter-label" for="filter-${filterType}" >
@@ -12,7 +12,7 @@ const createFilterItemTemplate = (filterType) => `
   </div>
 `;
 
-const createFiltersTemplate = () => `
+const createFiltersTemplate = () => /*html*/`
     <form class="trip-filters" action="#" method="get">
       ${FILTER_TYPES.map((filterType) => createFilterItemTemplate(filterType)).join('')}
       <button class="visually-hidden" type="submit">
